refactor(test): split kernel tests into focused modules
Extract shared recalcKernelEventId helper; exclude helper from build output.

Made-with: Cursor
